feat(restify): add requestHook support (#1312)
* feat(restify): add requestHook support

The `requestHook` config option allows custom span handling per request
layer.

* fix(restify): pass config to super class

As mentioned in the review, pass the instrumentation config to the
parent class. That way the config is also stored when given to the
initializer, rather only when using the `setConfig` function.

* fix(restify): fix comment referencing restify type

Update comment to reference to correct type from the `@types/restify`
package.

* fix(restify): import missing Span type

Add the missing import reported by the linter.

* fix(restify): fix issues reported by linter

* fix(restify): fix comment referencing restify type

Mention the package name exactly.

* fix(restify): fix comment referencing restify type

Co-authored-by: Amir Blum <amirgiraffe@gmail.com>

* fix(restify): add missing import in restify test

* feat(restify): add layer argument to requestHook

Add the layerType argument to the requestHook function. This is like the
following PR but for restify:
https://github.com/open-telemetry/opentelemetry-js-contrib/pull/1226

Move the LayerType from internal-types to types, because it's now used
in a function used by users of the instrumentation package.

Co-authored-by: Amir Blum <amirgiraffe@gmail.com>