fix(instrumentation-fastify): fix span attributes and avoid FSTDEP017 FastifyDeprecation warning for 404 request (#1763)
For a 404 `request.routeOptions.url` is undefined. Since fastify@4.10.0
when routeOptions was added, we shouldn't fallback to the deprecated
request.routerPath.

This also corrects the assumption that the handler name is "bound ..."
in all cases. E.g. for a 404 it is Fastify's core "basic404" internal
function.

Also add a test that Fastify instrumentation works for ESM usage.

Fixes: https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1757
Co-authored-by: Marc Pichler <marc.pichler@dynatrace.com>