fix(instrumentation-redis-4): fix multi.exec() instrumentation for redis >=4.6.12 (#1904)
In redis@4.6.12 the behaviour of multi.exec() changed to *throw* a
MultiErrorReply if any of the commands errored out. The individual
command replies are available at 'err.replies', instead of as the
promise result. This adjusts the instrumentation to generate
spans as before: only setting SpanStatusCode.ERROR and calling
span.recordException for the individual commands that failed.

Fixes: https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1874
Co-authored-by: Daniel Dyla <dyladan@users.noreply.github.com>