chore: add 'lint:deps' script to check for unused and unlisted deps (#2477)
This uses knip (https://knip.dev) to check dependencies, in hopes of avoiding
issues like #2473. This adds 'npm run lint:deps' and adds that to the
existing 'npm run lint'.

This change includes fixes for a handful of unused and unlisted deps.
For now knip is configured to only check 'production' deps. Checking non-prod
deps results in way too many false positives.

Note that knip is being run via `npx` rather than installing as a devDep because
there is a conflict: knip deps on typescript@5 and all packaegs in this repo
currently use typescript@4.4.4. Installing knip at the top-level results in a
ballooning of the node_modules install size as typescript@4.4.4 is installed
40+ types in all separate workspaces.